refactor: rename common/features/ApiResponse
to common/features/notifications
